AP_Logger: split methods from AP_LogFile into AP_Logger and AP_Logger_Backend
